Source-Linked Logs: record and link call location of debug messages in (HTML) log files
For someone unfamiliar with the source code, it is sometimes hard
to locate the code which produced a certain message. This patch
solves this issue like this:
- produce a fully cross-referenced HTML version of the source
  with Doxygen
- extend the logging functions and macros so that each message
  comes with file name, line number and function name information
- turn the time stamp in the HTML output into a link to the source
  code line, with the function name as title

The reason for picking the time stamp was that making other text a
link would interfere with the visual markup of it (color, bold, etc.).

In such an enriched log one can:
- hover over the time stamp to see function name in a bubble help
  and the URL (and thus source code location) at the bottom of the
  browser window (at least in Firefox)
- click on the time stamp to jump to the source code

Because this feature makes both the library and the logs larger,
it is turned off by default. Configure with --enable-debug-logs
to enable it when using autotools. Macro tricks are used to
produce the same code as before when the feature is off.

Limitations:
- Without time stamp logging, no links are recorded.
- Each change in the Synthesis source code forces a complete
  new run of Doxygen. This makes compilation quite a bit slower,
  but the only alternative would be to accept that logs and source dump
  are inconsistent in the changed parts.
- The links contain the absolute file path (making it a bit harder
  to share logs and source code dumps).
- The TClientEngineInterface::debugPuts() API introduced for and used
  by SyncEvolution leads to links which contain the wrong absolute
  path to the Doxygen HTML files (inside libsynthesis instead of
  SyncEvolution).
- The "prefix" parameter in that same API is still ignored.