time zone init and logging: split into part before and after config parsing
This patch reverts the "move system time zone parsing after config
reading" part of commit fc000b0f45. As Lukas pointed out, config parsing
already depends on time zones, for example for <definetimezone>.

But this patch keeps the finishConfig() callback and uses it to invoke
a second global time zone function. On Linux, that function is used
to add the libical time zone definitions and logs them at the same time.
Builtin time zones are added at the original time, before parsing the
config.

This is a compromise between having no time zones while parsing the config
and not logging the libical time zones. It works for SyncEvolution, because
it doesn't reference the additional time zones in the config.