Source-Linked Logs: fixed to work with older monolithic Synthesis targets as well
TDBG_LOCATION_xxx macros must be defined in sysync_debug.h
because in some C-only files, debuglogger.h cannot be
included.

Note that actually using SYDEBUG_LOCATION in a monolithic
build would still not work because SYDEBUG_LOCATION
relies on a constructor for a struct which does not
exist in plain C. But this does no harm as SYDEBUG_LOCATION
is not required in any of the classical builds.