dataobjtype: fixed bug that caused dataobjs without a profile (e.g. "file") to crash on creation.
For dataobjs, a profile is optional for parsing embedded objects,
like an email. For example FileObjs do not need a profile.

The case that no profile was specified, and hence no
configuration existed for it, was not handled, so
the code attempted to call newProfileHandler() on a
NULL pointer.
However, the case that no profile handler exists was already
properly checked for.