VTIMEZONEtoTZEntry/GetTZInfo: fixed unitialized read of bias
valgrind found an issue in the recent change to these functions:
if no DAYLIGHT saving was present in a VTIMEZONE, GetTZInfo() returned
true (because that is an acceptable VTIMEZONE), but dBias in
VTIMEZONEtoTZEntry() was not set either.

Perhaps a tri-state result in GetTZInfo would be needed: no
information, success, failure. Instead this patch changes GetTZInfo()
so that it does not modify its retvals when no information is found
(like it was done before the recent patch) and adapts
VTIMEZONEtoTZEntry() so that it deals with this (by setting
dBias in advance).