engine 3.4.0.14: implemented filtered sync set deletion
Before now, when part of the sync set was excluded locally
by filter expressions, zapping the datastore did still
delete the entire (unrestricted) sync set in case the
database was not able to apply the filters on the
database level. Although plugins can do that in theory,
no implementation exists for this so far. Only ODBC
based datastores have the capability to pass down
filter expressions into the SQL fetching statement.

To prevent wiping out more than the filter-defined
sync set also in cases the DB layer cannot or does not
filter, the one-by-one deletion routine in customimplds
now checks the need for filtering and applies
filters when needed (of course, this will cause the
DB to fetch all items to inspect them).

Furthermore, the plugin API DeleteSyncSet() is only called
when there are no filters or the plugin has acknowledged
trough FilterSupport() that it can fully handle
the filters. Otherwise, one-by-one deletion with
engine-side filtering is used as described above.