MIME Profile: check <parameter rule="something"> when generating
This patch adds new semantic: a parameter definition is ignored
during generation of an item unless no rule was specified or that
rule is active.

This semantic is simpler than the one for properties, which also
has the concept of a set of alternative properties of which only
one is selected.