MIME Profile: check <parameter rule="something"> when parsing
This patch adds new semantic: a parameter definition is ignored
while parsing an item unless no rule was specified or that
rule is active.

This semantic is simpler than the one for properties, which also
has the concept of a set of alternative properties of which only
one is selected.

The original use case for <parameter rule> only requires suppressing
the generation of parameters not understood by a peer. This patch here
was merely added to keep the functionality symmetric.