MIME Profile: setting one rule also activates all included rules
The MAKE/PARSETEXTWITHPROFILE() macro calls can only active one
rule. Allowing that rule to have <include> elements for other
rules *and* activating them together with the main rule allows
the macros to enables multiple rules.

Note that this activation is not recursive: subrules included by
subrules are not activated - perhaps they should be? This patch
follows the code in TSyncSession::checkRemoteSpecifics() which also
only activates the <remoterule> and the rules included in it,
without recursing.