engine 3.4.0.17: fixed bad timezones bug causing imported TZ to be treated as "Z" zone
Imported TZ get "CUR" in dynYear, but matchTZ was only looking for
entries with dynYear empty. The last TZ which has empty dynyear before
the dynamically added entries start is "Z".

So if a sync first imported a new TZ, and then used the same
TZ in subsequent records again, the first occurrence worked ok,
but the subsequent ones wrongly returned "Z" instead of the
imported zone.

This bug only shows when the remote VTIMEZONE does not match
one of the built-in zones (or one of those loaded at engine
startup or via config).