engine 3.4.0.20 - no longer forces slow sync on other profiles when one profile is zapped (with separate changelogs)
With combined changelog, it is inevitable to make the other profiles do a slow sync to
keep consistency (so this patch does not change that).

Also, with separate changelogs, we can keep syncs in other profiles resumable without
restrictions, so cancelling resumeAlertCode is not done any more with separate
changelogs (but is still required for combined changelog).