SyncML Toolkit: fixed memory leak in case of error
When accessing a SyncML server with invalid credentials, the session
is aborted without ever going through the normal decoder shutdown code
in mgrProcessNextCommand(). In that case xltDecTerminate() is not
called. Instead freeInstanceInfo() incorrectly calls smlLibFree() on
the decoder pointer, which leaks all decoder memory allocated in
addition to the context struct itself, as reported by valgrind.

A similar mistake was made for the encoder pointer, although that
hasn't shown up in valgrind yet.

This patch replaces smlLibFree() with xltDecReset()
resp. xltEncReset().