temporary IDs: avoid reusing existing ID
I see failures where the server assigns the same temporary ID to
different items in the same sync.

I've added debug logging. It shows that the following happens:
     1. fTempGUIDMap is restored from the map file such that it has 105
        entries, *but* these are non-contiguous (from #1 to #124).
     2. fTempGUIDMap.size() + 1 is #106, which already exists in
        fTempGUIDMap.
     3. Overwriting that entry does not increase fTempGUIDMap.size().
     4. A second item is assigned the same #106 temp ID.

I don't know exactly how I arrived at this state and whether it is valid.

This patch works around the problem by checking for an ID conflict and
iterating until it finds a truely unused ID, instead of assuming that
"hash size + 1" yields a new ID.

The debug logging is also part of the patch, in case that further
analysis is needed.