maximum UID size: avoid redundant tempUID->localUID mappings
The storage of map items maps from (local id, ident) to ([remote id],
flags) and thus cannot store multiple remote ids for the same local
id. fTempGUIDMap ended up having such duplicate entries if:
- there was an entry from a previous updating of item on the client
- the item was deleted on the server, creating a new temporary id
  to report deletion to the client

Such a fTempGUIDMap cannot be stored and retrieved again without
loosing the older temporary mappings, which a) might have been still
needed (not sure) and b) violated the assumption that temporary ids
have to be sequential.

This patch checks for an older mapping the the same local id and
reuses that mapping instead of creating a new one. It uses a simple
linear search, because that is all that can be done unless a more
complex data structure is introduced (which might not be worth the
effort).