vtimezone: retain original TZID in imported definitions
VTIMEZONEtoInternal() used to overwrite the original TZID because
there was no other place to store standard and daylight saving TZNAME.
Since there are now explicit fields for that, the code overwriting the
TZID can be removed.

The advantage is that importing and exporting a custom definition
introduces less changes, because now the original TZID will be exported.