vCalendar 1.0: avoid RRULE end date without time
When parsing an iCalendar 2.0 event, it is possible to arrive at
an rrule "until" date/time which contains no time value. vCalendar 1.0
requires such a time. Not sending it, as the engine did before, causes
Nokia phones to ignore the recurrence rule (bugs.meego.com #11241).

This patch avoids this case by looking at the time of the event start
and adds that time to the recurrence until date. The comment in the
default config says:
      <!-- recurrence rule block, fields must be in that order, including
           DTSTART as last field !! -->

But it doesn't seem like DTSTART was really used before. Therefore this
patch tries to deal with not finding a timestamp after the recurrence rule
block by clearing the TCTX_DATEONLY, without actually setting a time. Probably
not quite correct, but perhaps better than not changing anything.

Another problematic aspect is to match the time zone context of the
"until" clause with the "start" date/time of the event. Currently the patch
assumes that the start time is going to be encoded as UTC.