engine: server: optimized tempGUIDs - only generate them when needed.
Before, all items that had a localID (=all) were run through adjustLocalIDforSize(),
so even if <alwayssendlocalid> was configured to off, even replace and delete
items received a tempGUID, which then was removed in newSyncOpCommand() again.

This is rather inefficient as it causes map items to be generated and even made
persistent in the DB that are never used at all.

The solution was to move both tempGUID checking/generating and ID prefixing
(needed for superdatastores) into newSyncOpCommand() after the point where
the decision for including a localID or not is made.