SyncML TK: adjust size of DevInf buffer dynamically
The size of the buffer into which DevInf was encoded was limited to
~12KB, regardless of which message size was configured for the actual
message.

12KB is too small for larger numbers of databases, like 15, which
caused "SmlException: SyncML Toolkit error=0x2002, smlResultsCmd,
smlerr=8194".

This patch uses a DevInf buffer size which is as large as the
remaining buffer for the actual message, or the old size (just in
case, not expected to be used anymore).