binfile client: allow configs without server URL
TBinfileImplClient::SelectProfile() checked whether a server URL was
configured. Such a URL is not needed in some cases, for example when
using a pre-configured transport, as in SyncEvolution's local sync
(Synthesis client <-> Synthesis server communicating via pipes).

Setting a dummy URL would have worked around the issue, but is not
the right solution. Removed the check instead.