TMimeDirProfileHandler::getOptionsFromDatastore(): clarify usage of session pointer
It is a bit confusing that TMimeDirProfileHandler::getOptionsFromDatastore()
depends on fRelatedDatastoreP to get a session pointer when it also
could use its own getSession(). The reason is that when running as part
of a script, fRelatedDatastoreP is NULL and thus copying the wrong
options for the remote peer from the session is avoided.

This patch clarifies that and replaces the many calls to
fRelatedDatastoreP->getSession() with a local variable. It also checks
that the result really isn't NULL.