MAKE/PARSETEXTWITHPROFILE(): allow rule parameter to influence global conversion options
So far the optional rule parameter was only used to select properties
in the profile via their "rule" parameter. In contrast to a sync session,
it wasn't possible to enable <noemptyproperties>, for example.

This patch makes that possible by extending the
TMimeDirProfileHandler::setRemoteRule() semantic: now each rule and
its sub-rules are also check for relevant conversion settings. Because
there is no session where these settings could be stored, some code had
to be copied from TSyncSession::checkRemoteSpecifics(). For the sake of
completeness, all settings which do not apply are still in the patch
as lines that were commented out.

Because setRemoteRule() is not used outside of these two script macros
(clarified in a comment), this change has no negative effect on
syncing.