mime-dir: Improved QP decoding, now detects not only CRLF, but also CR and LF within QP. Also fixed generator to avoid softbreak indicator at end of string.
Although content linefeeds should be CRLF (=0D=0A when QP encoded), Funambol
seems to use LF only (=0A).
Up to now, the QP decoder only recognized the CR and ignored LFs, so newlines
got lost with LF-only encoding.
Now the detector works like in other places of the engine, in that it
accepts CR, CRLF, and LF sequences as a newlines.