version bump: new revision, mark in .pc file
The VERSION set by AM_INIT_AUTOMAKE() was only used in the *.pc files
and has never been updated since the initial Linux release. Bumped it
to 3.4, so that SyncEvolution can test for a recent enough libsynthesis
during configuration.

Also incremented the .so. New features were added, without breaking
backwards compatibility. This is not quite the same as extending the
ABI, but close.