remote rules + OverrideDevInf: parse DevInf as part of checking the config
Now the DevInf is parsed in TSessionConfig::localResolve(). The result
is stored in the TRemoteRuleConfig instance.

The advantage is that syntax errors are found earlier and for all
remote rules, not just the one active during a sync. The downside is
somewhat increased memory and CPU consumption.

The DevInf in XML format is kept in memory although no longer needed
after localResolve(). That might be useful for debugging.