<comparescript>: added COMPAREMODE(), needed to determine compare mode
A compare script must know what the compare mode is. When checking for
"items equal", it must return 0 (equal) or -999 (not equal). When doing an
age comparison, it must 1 or -1.

The new COMPAREMODE() method returns a string which identifies the current
mode:
  "all",    // compare all fields, even irrelevant ones
  "n/a", // n/a (scripted)
  "conflict", // compare fields relevant for (normal sync) conflict detection
  "slowsync", // compare fields relevant for slow sync match
  "firstsync", // compare fields relevant for first time slow sync match (possibly relaxed comparison rules)
  "age", // no test for equality, only for age (returns SYSYNC_NOT_COMPARABLE if age comparison not possible)

A script only needs to check for "age", and it equal, do the age comparison
instead of checking for equality.