engine: updated implementation of "unprocessed" wildcard properties
The 3529d3cb commit which introduced support for "unprocessed" wildcard
properties added the logic that such properties are always considered
"available" regardless of the actual CtCap of the peer.

This has the undesirable effect that the content of these properties
are removed when receiving an update from a peer which doesn't
support them.

Therefore this commit reverts that part of 3529d3cb: setfieldoptions()
with NULL as first parameter now only enables mandatory properties,
as it did before the introduction of "unprocessed" properties.

"unprocessed" properties need to be listed explicitly in the CtCap
like any other property. This is currently untested.