server engine: better support for backend doing its own duplicate merging (status 207 from API)
So far, the DB backend could return DB_DataMerged (=207) for an add
to signal that the added data was merged with some pre-existing data
and thus the client should be updated with the merge result. This was
intended for merge with external data (augmenting entries with
lookups from other sources), but not yet for duplicate elimination.

This patch adds support for propagating merges that eliminate duplicates
in addition to merges that just add external data.

When a client sends a new item (new = client side ID not yet known
server side), and the DB backend returns status 207 for AddItem,
the following happens:

- current maps are searched for an item with the same localID as
  what the DB backend just returned as the "new" localID. If such
  an item is found, this means the add was completed by merging
  with an existing item.
- If so, this means that for this single localID, there are now
  two versions with different remoteIDs on the client. The item
  with the remoteID found in the map must be deleted, so a
  delete command is added to the list of changes to be sent to
  the client.
- It might also be that the item merged was just added to the
  server (and not yet known to the client), or had another
  change predating the merge. If so, the list of changes to be sent
  to the client will contain an add or a replace, resp.
  that must NOT propagate to the client, so it is removed
  from the list now.
- Finally, the result of the merge is fetched from the database
  and propagated to the client as a replace command.