bumped version to 3.4.0.16.1 and .so revision by one
The first four digits are from the upstream tag. The last digit needs
to be incremented each time relevant changes are released as part of
the Linux libsynthesis release in SyncEvolution. This is relevant
primarily for the PKG_CONFIG_CHECK(), which now can check that the
libsynthesis that is installed is recent enough.