DB_Conflict (409): do age comparison before merging, avoid unneeded changes
Instead of always assuming that the incoming item is meant to win,
look at the age of the two items first.

Also tell the code which updates the DB item whether it was changed at
all and skip the DB write if it is not needed. Note that statistics
are still wrong: a server add is counted as "item added" in all cases,
even if it is turned into an update or becomes a nop.