DB_Conflict (409): avoid sending unnecessary changes back to client
In server mode, when the merging after a 409 from the DB determines
that the client already has the latest data, sending it a Replace
command can be avoided.

This patch achieves that by introducing a new boolean
remoteHasLatestData which is false by default and only set when it is
certain that sending the update can be avoided. To minimize changes,
the DB_DataReplaced code path is left unchanged. It could be changed
to set the same boolean.

Not sure whether that will also avoid unnecessary changes sent when a
client detects a conflict.