DB_Conflict (409): added missing implementation in binfileimplds.cpp
When used as client, the active implementation of implProcessItem() is
the specialized one in binfileimplds.cpp, not the one in customimplds.cpp.
Only the latter had the 409 merge feature so far. Now the binfileimplds.cpp
version has it, as well.

I also had to move mergeWithDatabaseVersion() helper method up the hierarchy
into stdlogicds.cpp to make it accessible for binfileimplds.cpp.