plugin API: pass firstReadNextItem=true on first call
The firstReadNextItem parameter of ReadNextItem() was always
false. The plugin was implicitly expected to start at the beginning.

Now the parameter is "true" once for the first call. Plugins which
ignore the parameter continue to work as before. Plugins which check
it should reset their "list items state", as required by the API.

The main reason for changing this is the new "use plugin instance in
multiple SyncML sessions" feature: if enabled for the data store,
listing items with ReadNextItem() will be done multiple times and then
reseting the "list item state" is essential.