TCustomImplDS::implMarkOnlyUngeneratedForResume: fixed endless loop
The while() loop gets stuck if any entry in the map has the "delete"
member set, because then the code does a "continue" which doesn't move
the pointer beyond that entry.

Fixed by using a for() loop.

Found while experimenting with better 409 conflict handling in the
binfile client, where one artificial local ID is marked as deleted.