DB_Conflict (409): different implementation in binfileimplds.cpp
The previous approach to handling a 409 status for an <Add> (= item
conflicts with existing item) in the binfile client was to update that
existing item and then tell the server that its item maps to an
existing client item.

This leads to the situation where the server has two client IDs
mapping to the same item on the server and one server item with no
corresponding client item. Servers get confused by this. For example,
the Synthesis engine itself then sends a <Delete> with empty IDs to
the client in the next sync.

It also has the disadvantage that the client cannot ask the server to
delete the redundant item, because its <Delete> requests must include
a client ID, which the redundant item doesn't have.

Furthermore, the server was sent a redundant <Update> in the case that
it already had an item that was in sync in the client (= item didn't
have to be written in local storage).

Therefore this commit implements a different approach:
- the new server item which triggers the 409 is used to update
  the existing item, as before
- if and only if the local item gets modified, it will be sent as
  update for the older server item already mapped to it (there must
  be such an item, because the client must have told the
  server about that local item, and that server item is now
  out-dated)
- the new server item is mapped to a fake client ID which is
  marked as deleted; in the next sync, that mapping is used
  to delete the new server item

Once a second sync is done, client and server are in sync again, with
the latest data as determined by the client stored in the servers
older item and the newer item deleted.

Generating a fake client ID is a bit hacky at the moment. The code for
numeric IDs is entirely untested, the code for string IDs doesn't
check for (unlikely?!) collisions.