SyncML extensions: multiple cycles in the same session
SyncML limits the data exchange to "client sends changes, server sends
changes back". If the client detects that it has further changes for
the server while already in the second phase, it will have to wait for
the next SyncML session before it can send these changes. This leads
to a bad user experience, because the user expectation is that client
and server are in sync after a sync session.

Examples when this becomes a problem:
- 409 conflict detected in binfile client (tested with
  SyncEvolution Client::Sync::eds_event::testAddBothSides[Refresh]
  and a SyncEvolution server with file backend which doesn't
  detect duplicates based on UID)
- temporary error prevents applying a change, needs to be
  retried (untested at the moment)
- PBAB use case: advanced client does a sync with a subset
  of the data (vCard without PHOTO, while retrieving PHOTO data
  in the background), then sends updates with more data (modified/new
  PHOTOs) later on (also untested)

This commit introduces a new mode where SyncML client and server allow
multiple read/write cycles inside the same session. This extends the
SyncML 1.2 standard in the following way:
- in refresh-from-client mode, client and server always enter the
  map phase although the standard says that it can and must be skipped;
  the Synthesis engine already supported that non-standard behavior
  via the fCompleteFromClientOnly flag (aka <completefromclientonly>
  in the config and remote rules)
- the client, if it has more changes for the server and hasn't encountered
  problem, finishes the current cycle internally and then sends a new <Alert>
  in the next message to start the next cycle, instead of ending the map
  phase and package with a <Final/> tag
- the server starts at the beginning of a new SyncML session when it gets
  that <Alert>, instead of rejecting it as a protocol violation
- the first sync mode is done as set by the app; any following one
  falls back to incremental mode:
  - slow sync -> two-way sync or one-way if data direction was limited
  - refresh sync -> one-way sync in same direction
  - one-way sync -> do it again

To simplify testing, this new functionality is currently limited to
sync sessions with only a single active data store. The new mode has
to be enabled *on both sides* with the <completefromclientonly> config
option. When only enabled on the client side, the unexpected <Alert>
will be rejected by the serer. When only enabled on the server side,
the client might be surprised by the non-standard map phase behavior.

SyncEvolution enables this new mode unconditionally with a remote rule
when talking to itself. Ideally this should be replaced with a DevInf
extension, both for "SyncML engine supports the mode" and "data store
supports restarting".

Data stores must be able to deal with repeated cycles in the same
session. Typically this consists of reseting some state so that
running another sync cycle doesn't reuse obsolete information. The
binfile DS supports this now, but the underlying data store in the app
might not be ready for it yet. Currently the engine does not check
this.

The binfile client automatically recognizes the 409 conflict case and
enters a new sync cycle to get the server in sync with the client. In
addition, any client can be told to restart the sync by setting the
new "restartsync" session variable (PBAP use case). This has to happen
before the client considers the session completed, i.e. before
EndDataWrite() is called. It would be better to do this with a
per-datastore flag, but it wasn't obvious how to expose such a flag
via the engine API.

While looking at the code it was noticed that issueCustomEndPut() is
not called when running in standard-compliant refresh-from-client
mode, because the code is only invoked at the end of the map phase,
which is skipped in that mode. A bug?! Not fixed.