datastore: explicitly tell the engine whether restarting a sync is supported
The non-standard "multiple sync cycles in the same session" mode
depends on data stores which support multiple read/write cycles and
properly report changes at the start of additional cycles. The engine
will automatically run multiple cycles if it deems that necessary to
get client and server in sync.

This commit introduces the necessary flag that must be set for a store
before the engine can be sure that it is safe to run more than one
cycle. Not currently checked by the engine, though!