SyncML extensions: multiple cycles + DevInf
The PUT of DevInf in the first message was repeated in each cycle
because it wasn't checked whether the DevInf had already been
sent (is never true there in a normal SyncML session).

When restarting syncing, resending the DevInf is not useful. It is
uncertain how changes to the DevInf should be handled, so they are
simply not supported and sending the DevInf multiple times is
prevented by checking fRemoteGotDevinf.