restart sync: extend DevInf and check whether datastores support restarts
DevInf/Datastore/SyncCap is defined as a list of PCDATA strings.
Only the values for the standard DS sync modes are defined, other
values are explicitly allowed. We use that extension mechanism
to report additional flags for datastores.

The first such flag is "does the datastore support restarts", by
adding and checking for X-SYNTHESIS-RESTART in the SyncCap. Internally
the flag is now stored in SyncDataStore, where it is set when
receiving a suitable DevInf (remote datastore) or it was set in the
config (local datastore).

The logic which restarts a sync was adapted accordingly. Instead of
assuming that restarting is possible with the peer if
<completefromclientonly> was set for it, actually check the remote
datastore's "can restart" flag. The expectation is that this flag is
not set unless both datastore and SyncML engine support it, so there's
no need for a separate global flag.